Move notification listener to helpers. Rename error directory to errors
