+++ kiosk/src/js/application_updater.js
@@ -2,10 +2,10 @@ const app = require('electron').app;
 const autoUpdater = require("electron-updater").autoUpdater;
 const CurrentScreenSettingsManager = require('./current_screen_settings_manager');
 const cron = require('node-cron');
-const NotificationListener = require('./notification-listener/notification_listener');
+const NotificationListener = require('./helpers/notification_listener');
 const PropertiesLoader = require('./helpers/properties_load_helper');
 const UserInteractionsManager = require('./user_interactions_manager');
-const NetworkErrorsHandlingService = require('./services/error/network_errors_handling_service');
+const NetworkErrorsHandlingService = require('./services/errors/network_errors_handling_service');
 const AutoupdateScheduleWatcher = require('./services/data/autoupdate_schedule_watcher');
 const ConnectionStatusService = require('./services/network/connection_status_service');
 const Logger = require('./services/logger/logger');
