+++ kiosk/src/js/helpers/notification_listener.js
