+++ kiosk/src/js/services/errors/global_errors_handling_service.js
