+++ kiosk/src/js/services/errors/network_errors_handling_service.js
