+++ kiosk/src/js/services/errors/uncaught_errors_handling_service.js
