+++ kiosk/src/js/services/events/external-event-listeners/api_events_listener_service.js
@@ -1,4 +1,4 @@
-const NotificationListener = require('../../../notification-listener/notification_listener');
+const NotificationListener = require('../../../helpers/notification_listener');
 
 let notificationListener = new NotificationListener();
